Add DMP phases
